Update LoginUserControl.xaml.cs
I wasn't sure the best way to interact with the Wizard from inside of a usercontrol, so I opted to pass the object at init. I'd love to know a better way.

When the control is Loaded and ready, focus the Username box.

After login success, hide the popup window and go to the next screen automatically. I'm enabling the Finish button here because I added a new screen at the end.